# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "API Documentation",
        "contact": {
            "email": "ekoabdulgoffar129@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:49/bpm-backend-laravel/api/v1",
            "description": "API Documentation"
        }
    ],
    "paths": {
...
        "/api/v1/pertanyaan": {
            "get": {
                "tags": [
                    "Pertanyaan"
                ],
                "summary": "Get list of pertanyaan",
                "operationId": "e2275182d686546b516e9527de5bcca9",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "pty_id": {
                                                "type": "integer"
                                            },
                                            "pty_pertanyaan": {
                                                "type": "string"
                                            },
                                            "pty_status": {
                                                "type": "integer"
                                            },
                                            "pty_isheader": {
                                                "type": "integer"
                                            },
                                            "pty_isgeneral": {
                                                "type": "integer"
                                            },
                                            "pty_created_by": {
                                                "type": "string"
                                            },
                                            "pty_created_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "pty_modif_by": {
                                                "type": "string"
                                            },
                                            "pty_modif_date": {
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "ksr_id": {
                                                "type": "integer"
                                            },
                                            "pty_role_responden": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Pertanyaan"
                ],
                "summary": "Create new pertanyaan",
                "operationId": "9b30e320c6d792e29c2d5c94ac8f1f52",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pty_pertanyaan": {
                                        "type": "string"
                                    },
                                    "pty_status": {
                                        "type": "integer"
                                    },
                                    "pty_isheader": {
                                        "type": "integer"
                                    },
                                    "pty_isgeneral": {
                                        "type": "integer"
                                    },
                                    "pty_created_by": {
                                        "type": "string"
                                    },
                                    "pty_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "pty_modif_by": {
                                        "type": "string"
                                    },
                                    "pty_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ksr_id": {
                                        "type": "integer"
                                    },
                                    "pty_role_responden": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pertanyaan created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/pertanyaan/{id}": {
            "get": {
                "tags": [
                    "Pertanyaan"
                ],
                "summary": "Get pertanyaan by ID",
                "operationId": "9a7b8359123c2e31372a93c350fc1fe5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pertanyaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "pty_id": {
                                            "type": "integer"
                                        },
                                        "pty_pertanyaan": {
                                            "type": "string"
                                        },
                                        "pty_status": {
                                            "type": "integer"
                                        },
                                        "pty_isheader": {
                                            "type": "integer"
                                        },
                                        "pty_isgeneral": {
                                            "type": "integer"
                                        },
                                        "pty_created_by": {
                                            "type": "string"
                                        },
                                        "pty_created_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "pty_modif_by": {
                                            "type": "string"
                                        },
                                        "pty_modif_date": {
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "ksr_id": {
                                            "type": "integer"
                                        },
                                        "pty_role_responden": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pertanyaan not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Pertanyaan"
                ],
                "summary": "Update pertanyaan",
                "operationId": "3b8261b3806dc0fbdcc3dd8e36f919b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pertanyaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pty_pertanyaan": {
                                        "type": "string"
                                    },
                                    "pty_status": {
                                        "type": "integer"
                                    },
                                    "pty_isheader": {
                                        "type": "integer"
                                    },
                                    "pty_isgeneral": {
                                        "type": "integer"
                                    },
                                    "pty_created_by": {
                                        "type": "string"
                                    },
                                    "pty_created_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "pty_modif_by": {
                                        "type": "string"
                                    },
                                    "pty_modif_date": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "ksr_id": {
                                        "type": "integer"
                                    },
                                    "pty_role_responden": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pertanyaan updated successfully"
                    },
                    "404": {
                        "description": "Pertanyaan not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pertanyaan"
                ],
                "summary": "Delete pertanyaan",
                "operationId": "f15ed38c9526174c06ed1348a909f05c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID pertanyaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pertanyaan deleted successfully"
                    },
                    "404": {
                        "description": "Pertanyaan not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token",
                "scheme": "bearer"
            }
        }
    },
...
}